Move WikiLove preference from Misc. to Editing tab
Misc. is the worst tab ever in preferences and should be removed.
WikiLove has to do with Editing, so the Editing tab makes the most
sense to move it to.

Despite the name, editing/advancedediting is not a section called
"Advanced" but rather "General" so it seems to make sense as a new
home.

Bug: 52757
Change-Id: I77d79003b8b8f20ae2df24fe5e058dffb6a0399d