Remove PNG fallback for JS-only interface
This has been applied to OOUI, VE, WikiEditor and various other
places in past years, but not yet here. SVG is supported by all
JS-enabled browsers. The icon is hidden on client-nojs, so we
don't need a PNG fallback.

Also remove '@embed' as the rendering of the article text
should not be delayed by this icon.

Bug: T121730
Change-Id: I6abf4a3bf3ace90684b0554f0d2da6b98f6c689e