Add support for passing tags to api via $.wikiLove.openDialog
In order to add a page triage tag when sending wikilove is triggered
from the page triage interface, there needs to be a way to pass
the tag when starting the dialog, since the call to openDialog
is the only interaction between the page triage extension and the
wikilove extension. There is currently an optional recipients parameter,
and this patch adds a second optional parameter for tags to be applied
to the WikiLove action.

Bug: T231541
Change-Id: I9c4f7b78ecf84bc54729b64cfd1498cb8a65862d