Fix broken "WikiLove.js" default value
Follows-up cb7d900eb0f7c (I2b622c5ce72bc) from 2014 which meant to
change the casing of the "L" but also ended up changing the
initial letter.

It is possible that worked in MediaWiki back then, but it doesn't
today. Which means the helpful comment is not shown when accessing
this page initially, and it means Special:AllMessages shows the
override as replacing the default value of ⧼wikiLove.js⧽ instead
of the actually intended default value.

Change-Id: I19a58b10fa0a2ef6dabe1be65a145be29419af75