Set page title on preview, pass section header as is
Setting the page title in the preview makes sure that any page-dependent
things (in particular, DiscussionTools visual enhancements, but also for
example magic words if used) work as expected, at least on wikitext talk
pages. (On Flow boards, the final page is the thread page, while the
preview happens on the board itself, but I don’t think we can do
anything about it.)

Passing the section header makes sure that (again, on wikitext talk
pages) the correct heading level is used if it has been customized using
the `newsectionheaderdefaultlevel` MediaWiki message.

Bug: T321985
Change-Id: I8939f1f1c051e0f17f329d45e449356ade6371de