darkmode: Use color-tokens for the entire interface as needed
This does not tough the green circles with the numbers 1-3 in them. They
are not meeting the needed contrast, but that was already the case in
light-mode. So if we want to change them, maybe that should be a
dedicated commit.

Also this does not change the default styles in the table for those
messages that come with one, e.g. the food. This will need a bit of
extra work, because the table used to not include the font-color, but in
dark-mode that is necessary due to the decision found in T358240 (see
also ADR-0005 in the Minerva repo).

Bug: T375351
Change-Id: I07c0a606f7238fbf878da81e705b498aca58d2c0